+++ src/error.rs
@@ -3,6 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 use std::fmt;
+use std::str;
 
 /// Position of an error.
 ///
@@ -25,7 +26,7 @@ impl ErrorPos {
     }
 }
 
-impl fmt::Debug for ErrorPos {
+impl fmt::Display for ErrorPos {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}:{}", &self.row, &self.col)
     }
@@ -73,35 +74,50 @@ pub enum Error {
     },
     /// An SVG element must contain a tag name.
     ElementWithoutTagName(ErrorPos),
+    /// UTF-8 processing error.
+    Utf8Error(str::Utf8Error),
 }
 
-impl fmt::Debug for Error {
-
+impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
             Error::UnexpectedEndOfStream(ref pos) =>
-                write!(f, "Unexpected end of stream at {:?}", pos),
+                write!(f, "Unexpected end of stream at {}", pos),
             Error::InvalidChar{ref current, ref expected, ref pos} =>
-                write!(f, "Expected '{}', found '{}' at pos {:?}", expected, current, pos),
+                write!(f, "Expected '{}', found '{}' at pos {}", expected, current, pos),
             Error::InvalidSvgToken(ref pos) =>
-                write!(f, "Invalid SVG token at {:?}", pos),
+                write!(f, "Invalid SVG token at {}", pos),
             Error::UnexpectedClosingTag(ref pos) =>
-                write!(f, "The stream found closing tag without an opening tag at {:?}", pos),
+                write!(f, "The stream found closing tag without an opening tag at {}", pos),
             Error::InvalidNumber(ref pos) =>
-                write!(f, "Invalid number at {:?}", pos),
+                write!(f, "Invalid number at {}", pos),
             Error::InvalidLength(ref pos) =>
-                write!(f, "Invalid length at {:?}", pos),
+                write!(f, "Invalid length at {}", pos),
             Error::InvalidColor(ref pos) =>
-                write!(f, "Invalid color at {:?}", pos),
+                write!(f, "Invalid color at {}", pos),
             Error::InvalidTransform(ref pos) =>
-                write!(f, "Invalid transform at {:?}", pos),
+                write!(f, "Invalid transform at {}", pos),
             Error::InvalidAttributeValue(ref pos) =>
-                write!(f, "Invalid attribute at {:?}", pos),
+                write!(f, "Invalid attribute at {}", pos),
             Error::InvalidAdvance{ ref expected, ref total, ref pos } =>
-                write!(f, "Attempt to advance to the pos {} from {:?}, but total len is {}",
+                write!(f, "Attempt to advance to the pos {} from {}, but total len is {}",
                        expected, pos, total),
             Error::ElementWithoutTagName(ref pos) =>
-                write!(f, "An element without a tag name at {:?}", pos),
+                write!(f, "An element without a tag name at {}", pos),
+            Error::Utf8Error(e) =>
+                write!(f, "{}", e),
         }
+    }
+}
+
+impl fmt::Debug for Error {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", &self)
+    }
+}
+
+impl From<str::Utf8Error> for Error {
+    fn from(value: str::Utf8Error) -> Error {
+        Error::Utf8Error(value)
     }
 }